+++ src/ZoneTree/WAL/IWriteAheadLogProvider.cs
@@ -6,23 +6,23 @@ namespace ZoneTree.WAL;
 public interface IWriteAheadLogProvider
 {
     IWriteAheadLog<TKey, TValue> GetOrCreateWAL<TKey, TValue>(
-        int segmentId,
+        long segmentId,
         ISerializer<TKey> keySerializer,
         ISerializer<TValue> valueSerialize);
 
     IWriteAheadLog<TKey, TValue> GetOrCreateWAL<TKey, TValue>(
-        int segmentId,
+        long segmentId,
         string category,
         ISerializer<TKey> keySerializer,
         ISerializer<TValue> valueSerialize);
 
-    IWriteAheadLog<TKey, TValue> GetWAL<TKey, TValue>(int segmentId);
+    IWriteAheadLog<TKey, TValue> GetWAL<TKey, TValue>(long segmentId);
 
-    IWriteAheadLog<TKey, TValue> GetWAL<TKey, TValue>(int segmentId, string category);
+    IWriteAheadLog<TKey, TValue> GetWAL<TKey, TValue>(long segmentId, string category);
 
-    bool RemoveWAL(int segmentId);
+    bool RemoveWAL(long segmentId);
 
-    bool RemoveWAL(int segmentId, string category);
+    bool RemoveWAL(long segmentId, string category);
 
     void DropStore();
 }
