+++ src/ZoneTree/WAL/NullWriteAheadLogProvider.cs
@@ -5,32 +5,32 @@ namespace ZoneTree.WAL;
 
 public class NullWriteAheadLogProvider : IWriteAheadLogProvider
 {
-    public IWriteAheadLog<TKey, TValue> GetOrCreateWAL<TKey, TValue>(int segmentId, ISerializer<TKey> keySerializer, ISerializer<TValue> valueSerialize)
+    public IWriteAheadLog<TKey, TValue> GetOrCreateWAL<TKey, TValue>(long segmentId, ISerializer<TKey> keySerializer, ISerializer<TValue> valueSerialize)
     {
         return new NullWriteAheadLog<TKey, TValue>();
     }
 
-    public IWriteAheadLog<TKey, TValue> GetOrCreateWAL<TKey, TValue>(int segmentId, string category, ISerializer<TKey> keySerializer, ISerializer<TValue> valueSerialize)
+    public IWriteAheadLog<TKey, TValue> GetOrCreateWAL<TKey, TValue>(long segmentId, string category, ISerializer<TKey> keySerializer, ISerializer<TValue> valueSerialize)
     {
         return new NullWriteAheadLog<TKey, TValue>();
     }
 
-    public IWriteAheadLog<TKey, TValue> GetWAL<TKey, TValue>(int segmentId)
+    public IWriteAheadLog<TKey, TValue> GetWAL<TKey, TValue>(long segmentId)
     {
         return new NullWriteAheadLog<TKey, TValue>();
     }
 
-    public IWriteAheadLog<TKey, TValue> GetWAL<TKey, TValue>(int segmentId, string category)
+    public IWriteAheadLog<TKey, TValue> GetWAL<TKey, TValue>(long segmentId, string category)
     {
         return new NullWriteAheadLog<TKey, TValue>();
     }
 
-    public bool RemoveWAL(int segmentId)
+    public bool RemoveWAL(long segmentId)
     {
         return false;
     }
 
-    public bool RemoveWAL(int segmentId, string category)
+    public bool RemoveWAL(long segmentId, string category)
     {
         return false;
     }
